GeraMapas

Description

Esta primeira de três tarefas permite a criação de um tabuleiro de dimensão e variada jogabilidade definida pelo utilizador. Usa funções recursivas definidas por nós, mas também algumas ja existentes no Haskell.

Documentation

mapa :: Int -> Int -> [String]

# -> representa pedra

? -> representa tijolo

'' -> representa célula vazia

+ representa bombs

! representa flames

x-coluna y-linha

Para fim de testes, usemos putStr (unlines (mapa n s)) no GHCi para visualizar o resultado.

auxMapa2 :: [String] -> [Int] -> [String]

A função auxMapa2 serve de auxiliar á função coordenadas Esta coloca os valores da lista Random no mapa.

rnd será a lista dos numeros random

fazemos insEsp x porque como (x:xs) = [String], o "x" será um [Char] (uma linha)

ret_Esp_Mapa :: [String] -> [String]

ret_Esp_Mapa --> esconde os powerUps que se encontram no mapa, isto porque estes estão sempre atrás de um bloco de tijolos ?

map (x -> if (x==+ || x==!) then ? else x) (x:xs) --> uma função que, se verificar x==+ ou x==! substitui esse x por um ?

A razão pela qual tivemos que fazer um debug desta função, foi porque nós precisavamos visualizar o resultado com os powerUps visíveis para calcular as coordenadas manualmente.

contagem :: [Char] -> Int

Conta o número de posições onde deverão ser colocados os valores do rnd, isto para depois retirar deste os valores que já foram colocados.

linhas :: Int -> Int -> [String]

Gerador do mapa com apenas as paredes (#). Nos espaços onde será colocado um nº random, estará um r para mais tarde substituir

linhas --> cria a lista com as linhas do mapa.

(replicate (n-6) r) --> apenas usada na primeira linha (linha 0 no Haskell) e na penultima linha (linha (n-1) no Haskell porque estas precisam de ter as linha 1 e (n-2) com dois lugares vazios no inicio e fim de cada respetiva linha, para o player conseguir se mover no inicio do jogo.

A linha(n-1) é a ultima linha Terá sempre (n-1) linhas pois a linha 1 equivale á posição real 0

A razão pela qual definimos dois casos especiais para quando n=5, foi porque este é o caso mais pequeno possivel e o método de criação das linhas exige que na linha1 e linha(n-2) (segunda e penúltima respetivamente) tenham sempre "6 espaços" já ocupados para podermos comparar com a dimensão, para ver quantas vezes repetimos o caracter r (para futura substituição por um nº do rnd) nestas linhas. Se n=5, como n-6=-1, que por sua vez -1<0, a função iria adicionar "-1" vezes r o que causa um erro.

coordenadas :: [String] -> [String]

Junta, com o auxilo da aux (B/F) as coordenadas de cada um dos caracteres para isto, esta usa a função zip já predefinida no Haskell em cada lista de caracteres para criar uma lista de tuplos (coordenadas de todos os caracteres)

Funçao que faz uma lista das linhas onde se encontram os powerups --> linhaPower
Depois uma funçao que da a lista das colunas onde os powerups se encontram --> colunaPower

contAux serve para contar o número de vezes que um determinado powerUp se encontra numa linha pois pode haver mais que um powerUp na mesma linha, assim como na coluna (auxcoluna)

auxB :: [(Int, Int)] -> [String]

auxF :: [(Int, Int)] -> [String]

linhaPower :: Int -> Char -> [String] -> [Int]

contAux :: Char -> [Char] -> Int

colunaPower :: Char -> [String] -> [Int]

auxcoluna :: Int -> Char -> String -> [Int]

mapaDebug :: Int -> Int -> [String]

Debug Area

Esta parte do código daqui pra frente serve para resolver bugs existentes no código.

As funçoes fazem quase o mesmo que as suas correspondentes, apenas possuem uma leves alteraçoes:

mapaDebug --> tem o mesmo efeito da função mapa, mas em vez de apresentar o resultado com os powerUps escondidos, mostra onde estes estão realmente. Esta funçao serviu para determinar se esta estava a produzir as coordenadas corretas.