Main

Description

Esta última tarefa executa uma compressão do jogo (a decorrer ou não), tornando-a mais compacta para armazenamento. Executa também a descompressão deste, retornando o estado do jogo tal como antes da compresso. Contém funções definidas por nos, mas também algumas já existentes no Haskell.

Documentation

main :: IO ()

calcdimensao :: [String] -> Int

Trata-se de uma função que recebe um mapa, nomeadamente o que está a ser utilizado e calcula a sua dimensão a partir do primeiro elemento da lista, que se trata da primeira linha.

Explicação da função:

Recebe uma [String] que se trata do mapa em utilização.

Recorre á função length e aplica-a á cabeça da lista pois para determinar a dimensão apenas necessitamos de um elemento da [String] (uma linha)

>>> ["###############","#    ?   ? ?  #","# # #?#?# #?# #","#?  ? ?       #","#?# #?#?#?# # #","# ?   ??      #","# # # #?#?#?#?#","#  ??     ? ? #","# # #?# # #?# #","#???       ?  #","# # #?# #?# #?#","#  ?     ?  ? #","# #?# #?# #?# #","#       ??    #","###############","+ 12 7","! 7 12","! 8 13","! 9 13","* 7 4 1 1 10","3 7 4","0 2 1"]
 = 15  

encode :: [String] -> String

Trata-se de uma função que comprime o estado atual do jogo. Devido a existirem caratéres que são imõveis (como as paredes a.k.a '#') estes são instantanêamente eliminados do mapa. Os espaços e os carateres especiais são substiuídos pelo caratér p. É de referencia também que pretendemos manter as coordenadas do mapa tal como estão, logo foi feita uma condição que diz que se o ultimo elemento de um dos elemntos da [String] não for igual a '#' então ele vai devolver esse elemento e aplicar novamente a funçao recursivamente, separando as coordenadas umas das outras com uma barra.

Explicação da função:

É fornecido à função encode o mapa em causa, de seguida, é utilizada a funçao "calcdimensao" desse mesmo mapa e depois a função "show" para que o número que a função calcdimensao nos fornece seja transformado em String.

De seguida adiciona uma '|' para que seja separado o valor da dimensao do mapa comprimido.

Logo após á aplicação do número da dimensão e da barra, a função recorre a uma função auxiliar "auxencode1".

A função "auxencode1" verifica se os elementos da [String] constituem o mapa em si, comparando o último elemento desse elemento da [String] a um "#", se este for igual então é aplicada a função "auxencode" que examina todos os elementos da String que pertence à [String]. Se este for "#" então são eliminados, se for igual a ' ' então é substituído por um char = e, se não verificar nenhum destes casos então será substituído por um char = c.

Se na função "auxencode1" não for verificada a condição de o último elemento de um elemento da [String] então a função vai devolver o elemento a ser avaliado pela função e adicionar uma "|" e após isto, fazer novamente a função "auxencode1".

Foi feita esta separação pois queríamos manter as coordenadas do jogo tal e qual como estavam na [String].

auxencode2 :: String -> Int -> String

auxencode1 :: [String] -> String

auxencode :: String -> String

decode :: String -> [String]

Pegando no resultado de encode, tranforma essa String do mapa num mapa igual ao original (antes do encode)

Syntax --> decode mapa = . . .

expandir :: String -> String

Quando encontra um nº, replica esse nº de vezes e

Se encontrar um c, não faz nada

Syntax --> expandir (encoded mapa) = . . .

fromDigInt :: String -> String

Na nossa função encode, no inicio da nova String tem um número e logo a seguir um "|" (para separar a dimensão dos demais caracteres)

fromDig_Int dá como resultado uma String com o número

Nota

Existe uma função no ghci que apresenta um número presente numa String na forma de Inteiro: read ()::Int, mas apenas pode ter números dentro dessa String, senão dá erro (no nosso caso, a String tem números e caracteres)

Syntax --> fromDigInt (encoded mapa)

linhas :: Int -> Int -> [String]

Função com a base da função linhas da "Tarefa 1", mas com uma mudança nos primeiros caracteres da primeira, segunda, penultima e última linhas: para as funções, ao usarem o resultado desta função, nao usarem os espaços especiais" (onde não pode ter nada), o resuldao ada linhas será uma [String], onde, por exemplo:

>>> linha 1 --> "#nnrrrnn#" , n==espaço especial       r==espaço a ser substituido

Nota

Temos um caso especial para quando n==5, pelos motivos apresentados na Tarefa 1

Syntax --> linhas n 0 = . . .

juntaParte :: String -> String -> String

Pegando numa linha do mapa e na string do encode, substitui cada caracter com "r" ou "n" da linha por uma letra da string do encode (e=" " && c="?" )

Exemplo

>>> juntaParte "#  rrrrrrrrr  # "eeeeepeeeepeeeepeeeepeepeeepeppe" = "#  eeeeepeee  #"

Syntax --> juntaParte list = . . .

juntaTotal :: [String] -> String -> [String]

Faz o mesmo que juntaParte, mas aplicado a um mapa ([String]) usando a juntaParte

>>> Exemplo: juntaTotal ["#  rrrrrrrrr  #","#r#r#r#r#r#r#r#"] "eeeeepeeeepeeeepeeeepeepeeepep" = ["#  eeeeepeee  #","#e#p#e#e#e#e#e#"]

Syntax --> juntaTotal list (encoded mapa) = . . .

mapa_modInicial :: String -> String

Esta e a próxima têm efeitos parecidos, apenas com algumas mudanças devidos a condições diferentes. Ambas apresentam uma String sem tudo o que está até a um "|".

mapa_modIntermedio :: String -> String

contagem :: String -> Int

Conta o nº de caracteres para serem substituidos existentes naquela linha para depois retirar esse nº de caracteres da String do encode

Serão "r" e "n"

coordsDownParte :: String -> String

Pega na String do encode e apresenta apenas a parte desta com as coordenadas de powerups, bombas e players

>>> = "|+ 3 3|! 5 5|* 7 7 1 1 10|0 4 3 +|1 7 7"

coordsDownTotal :: String -> [String]

Pegando no resultado de coordsDownParte e dá como resultado uma lista de String com cada coordenada

ex :: [[Char]]

Exemplo:

Mapa: ["######## ",""," ? ? ? ?","","#########","+ 3 3","! 5 5","* 7 7 1 1 10","0 4 3 +","1 7 7"]

length (concat mapa) = 115

encode mapa = "9|eeeeeeeeceeeeeeececeececeeceeecceeecceee|+ 3 3|! 5 5|* 7 7 1 1 10|0 4 3 +|1 7 7"

length (encode mapa) = 81

Redução = ((81 / 115) * 100) - 100 = 30% (aproximadamente)

Mais exemplos de teste

ex2 :: [[Char]]

ex3 :: [[Char]]