Tarefa6_li1g158

Documentation

type Coord = (Int, Int)

bot :: [String] -> Int -> Int -> Maybe Char

moves_final :: [String] -> Int -> Maybe Char

Avalia a posição do jogador e se as suas coordenadas pertencerem as coordenadas que vão ser afetadas pela bomba no mapa então é chamada a função descobre_proximo_passo. Caso contrário são comparadas as sua coordenadas á dimensão do mapa e atribuídos comandos para iniciar movimento.

moves :: [String] -> Int -> Maybe Char

Função que decide se irá ser utilizada uma função de jogada mais defensiva (descobre_prox_passo) ou uma função mais ofensiva (ofensivo) na descoberta do próximo passo do bot.

descobre_prox_passo :: [String] -> Coord -> [Coord] -> [Coord] -> Maybe Char

Função que trata de descobrir qual vai ser o próximo passo que o jogador vai efetuar. Verifica se certas coordenadas pertencem a coordenadas de bombas, ou bombas que vao explodir, ou se são paredes, e dá o comando ao bot. A função pode ser descrita como um sistema defensivo onde o jogador apenas "foge" aos problemas.

ofensivo :: [String] -> Coord -> Maybe Char

posBombas :: [String] -> [Coord]

Função que nos dá a posição de todas as bombas em coordenadas no mapa. Recebe um estado do jogo e devolve as coordenadas em que se encontram as bombas..

posPlayer :: [String] -> Int -> Coord

Função que nos dá a posição de um jogador em coordenadas no mapa. Recebe um estado do jogo e devolve a coordenada em que se encontra o jogador.

ex :: [[Char]]