Main

Description

Esta tarefa permite que o temporizador das bombas cheguem a zero, explodindo, assim como o "efeito espiral" no mapa quando o tempo chega a um determinado valor.

Documentation

type Mapa = [String]

type Coord = (Int, Int)

organiza :: [String] -> ([String], [String], [String], [String])

Funções responsáveis para organizar os tipos de dados internos, facilitando assim o seu desenvolvimento.

--------------------------------------------PARSE-----------------------------------

Transforma o tipo de dados externo para tipo interno (escolhido por nós).

so_mapa :: [String] -> [String]

A função so_mapa é aplicada a um mapa inteiro com PowerUps, jogadores e bombas mas apenas devolve a parte do mapa onde decorre o jogo sem informações adicionais. Função recursiva que quando vê uma String que não começa por '#' elimina-a.

so_powers :: [String] -> [String]

Função que recebe um mapa e devolve apenas as informaçôes que dizem respeito a PowerUps.

procura_bomba :: [String] -> [String]

Função que recebe um mapa e devolve apenas as informaçôes que dizem respeito a Bombas.

so_players :: [String] -> [String]

Função que recebe um mapa e devolve apenas as informaçôes que dizem respeito a Jogadores.

main :: IO ()

Secção da main da tarefa.

--------------------------------------------------------------------------------

avanca :: [String] -> Int -> [String]

Função avanca que permite a passagem do tempo.

mapa_final :: [String] -> [String]

Função que inicia as mudanças no mapa (chama outras funçoes)

auxiliar3 :: ([String], [String], [String], [String]) -> [String]

Aplica uma série de funções, específicas a determinadas partes do mapa, mas apenas existir bombas prestes a explodir (contador 0)

Casos:

>>> __muda3__ -> recebendo as __coordenadas afetadas pelas bombas__ e as __coordenadas das bombas__
                 verifica se alguma bomba que __explode__ afetará alguma /bomba existente/. Se isto
                 se verificar, esta bomba afetada terá o seu contador a __1__ (para explodir
                 no próximo /tick/.

>>> __muda2__ -> recebendo as __coordenadas afetadas pelas bombas__ e as __cordenadas dos players__
                 verifica se alguma bomba /afetará/ algum player. Se se verificar, este desaparece.

muda3 :: [(Int, Int)] -> [String] -> [String]

muda2 :: [(Int, Int)] -> [String] -> [String]

explode :: [String] -> [String]

apresenta a string com contador 0

muda :: [String] -> [String] -> [String] -> [String]

retira do mapa paredes e power ups afetados pela bomba que explodiu

compara_retira :: [String] -> [String] -> [(Int, Int)] -> [String]

compara e retira paredes, etc do mapa

pertence :: (Int, Int) -> [String] -> Bool

verifica se um par de coordenadas correspondem a alguma string do mapa (powerups,bombas ou players)

coords_afetadas :: [String] -> [(Int, Int)] -> [(Int, Int)]

Apresenta uma lista de coordenadas a serem afetadas pela bomba

aux_coords_afetadas :: String -> [(Int, Int)] -> [(Int, Int)]

auxiliar de coods_afetadas, onde é tratada uma bomba

aux_auxL :: (Int, Int) -> [(Int, Int)] -> Int -> Int -> [(Int, Int)]

Função que determina as coordenadas de uma bomba que irá explodir, com raio=r

Para isto, usa um contador k que serve para comparação com r. Quando k==r, a função termina.

aux_auxR :: (Int, Int) -> [(Int, Int)] -> Int -> Int -> [(Int, Int)]

aux_auxU :: (Int, Int) -> [(Int, Int)] -> Int -> Int -> [(Int, Int)]

aux_auxD :: (Int, Int) -> [(Int, Int)] -> Int -> Int -> [(Int, Int)]

coords_blocos :: [String] -> [(Int, Int)]

Determina todas as coordenadas referentes aos blocos '#' e ?

aux_blocos1 :: [String] -> Int -> [(Int, Int)]

aux_blocos2 :: String -> Int -> Int -> [(Int, Int)]

retirar :: [String] -> [String]

retira do mapa uma string

deletar :: (Int, Int) -> [String] -> [String]

retira do mapa uma string (outro tipo)

bombas_zero :: [String] -> Bool

verifica se existem bombas com contador "0"

auxiliar1 :: ([String], [String], [String], [String]) -> ([String], [String], [String], [String])

Executa a primeira mudança no mapa, neste caso passará o contador das bombas para tmp-1, através da muda_bomba

muda_bomba :: [String] -> [String]

Retira 1 segundo ao contador de todas as bombas existentes no mapa

stringToCoord :: String -> Coord

Função que recebe uma String que corresponde a informações de jogadores, bombas e powerUps e as transforma em coordenadas. Primeiro geramos a partir da String de cada elemento do jogo uma lista de Srtrings utilizando a função words . A partir da função words retiramos apenas o que nos interessa para gerar as coordenadas, os elementos da posição 1 e 2. Após a função words utilizamos a função drop e a função take.

numero :: String -> Int

A função número gera números inteiros (coordenadas) a partir de Strings, se a string tiver length 1, ou seja, for apenas um digito então é aplicada diretamente a fução digitToInt, caso contrário, se a String tiver length superior a 1 então utilizamos a função read para ler o número com mais de um algarismo.

direita :: Coord -> Int -> [Coord]

Estas quatro funções são utilizadas para gerar coordenadas que mais tarde possam ser compactadas em uma lista formando assim um ciclo da espiral. O processo foi dividido em 4 funçôes: direita pois é o primeiro passo da espiral, de seguida baixo, esquerda e finalmente cima. As funções possuem um contador para saber quando parar a sequencia.

baixo :: Coord -> Int -> [Coord]

esquerda :: Coord -> Int -> [Coord]

cima :: Coord -> Int -> [Coord]

faseEspiral :: Coord -> Int -> [Coord]

Função que recebe a coordenada inicial do início da espiral e a sua dimensão. A partir destes argumentos compacta em uma lista de Coordenadas um ciclo completo da espiral, chamando funções como direita , baixo , esquerda , cima. O mecanismo desta funçao consiste em útilizar o último elemento que foi gerado da função anterior para que seja criado um ciclo da espiral. O contador presente é essencial para se poderem realizar os casos de paragem.

coordenadasEspiral :: Coord -> Int -> [Coord]

A função coordenadasEspiral recebe a coordenada inicial de todas as espirais, (1,1), e a dimensão do mapa em causa. De seguida utiliza a função faseEspiral que vai gerar camadas da espiral sequencialmente formando "quadrados" quando dá uma volta inteira. Após uma volta a funçao recursivamente repete-se adicionando ás coordenadas iniciais mais um valor para iniciar um novo ciclo.

substitui :: Mapa -> Coord -> Mapa

A função substitui recebe o mapa em causa e a coordenada que vai ter de substituir pelo char '#'. A função é dividida em 4 partes: nv_mapa que utiliza a função so_mapa para que as alterações sejam feitas apenas na parte "jogável" do mapa; nv_Pu que utiliza a função so_powers para aplicar a função verifica apenas nas informações dos PowerUps; nv_Bombas que utiliza a função procura_bomba para aplicar a função verifica apenas nas informações das Bombas; nv_Pl que utiliza a função so_players para aplicar a função verifica apenas nas informaçôes dos Jogadores;

É utilizada a função splitAt na nv_mapa para separar a linha que pretendemos do mapa, e depois separar a coordenada que queremos substituir do resto da linha.

verifica :: [String] -> Coord -> [String]

A função verifica é utilizada nas listas de Strings que contêm as informações que dizem respeito aos PowerUps, Bombas e Jogadores. A função converte as strings em coordenadas para que possam ser comparadas com a coordenada que a espiral vai afetar. Se as coordenadas forem iguais, então a String é eliminada do mapa, caso contrário ela continua no mapa.

ex0 :: [[Char]]

Conjunto de exemplos de teste

ex1 :: [[Char]]

ex2 :: [[Char]]

ex3 :: [[Char]]

ex4 :: [[Char]]

ex5 :: [[Char]]

ex6 :: [[Char]]

ex7 :: [[Char]]

ex8 :: [[Char]]

ex9 :: [[Char]]

ex10 :: [[Char]]