Main

Description

Nesta tarefa, o jogo encontra-se funcional, com interface gráfica (graças ao Gloss), usando também as outras tarefas para implementar a passagem do tempo, movimentos e geração de mapas.

Documentation

data Bomba

Uma representação do estado do jogo.

Constructors

B

Fields

playerB :: Int

posB :: (Int, Int)

r :: Int

tmp :: Int

data Player

Constructors

Pl

Fields

player :: Int

pospl :: (Int, Int)

pM :: Int

pF :: Int

data PowerMais

Constructors

PM

Fields

posM :: (Int, Int)

data PowerFlame

Constructors

PF

Fields

posF :: (Int, Int)

type Estado = (Float, Mapa, [PowerMais], [PowerFlame], [Bomba], [Player], [Picture], Int)

parse :: [String] -> Estado -> Estado

Estado == (tmp_ini,m,pM,pF,b,pl,iM,time)

tmp_ini = variável usada para contar o tempo

m = mapa ([string] com apenas os blocos existentes no mapa)

pM = lista de powerMais

pF = lista de powerFlame

b = lista de bombas

pl = lista de players

iM = lista dos bmp usados para fazer os desenhos

time = contador com o tempo

Esta função permite transformar os dados recebidos num tipo de dados interno para facilitar o desenvolvimento.

>>> [String] ---> Estado

converteBomba :: String -> Bomba

Converte uma String de bomba para o tipo Bomba

convertePM :: String -> PowerMais

Converte uma String de PowerMais para o tipo PowerMais

convertePF :: String -> PowerFlame

Converte uma String de PowerFlame para o tipo PowerFlame

convertePl :: String -> Player

Converte uma String de Player para o tipo Player

contagemchar :: String -> Char -> Int

Conta quantos caracteres chr existem dentro de uma String

inv_parse :: Estado -> [String]

Converte do tipo de dados interno Estado para um tipo de dados que umas funções usam [String]

inv_convertepM :: [PowerMais] -> [String]

Converte dados do tipo PowerMais para String

Exemplo

>>> PM (x,y) = "+ x y"

inv_convertepF :: [PowerFlame] -> [String]

Converte dados do tipo PowerFlame para String

Exemplo

>>> PF (x,y) = "! x y"

inv_converteB :: [Bomba] -> [String]

Converte dados do tipo Bomba para String

Exemplo

>>> B pl (x,y) r tmp = "* x y pl r tmp"

inv_convertepl :: [Player] -> [String]

Converte dados do tipo Player para String

Exemplo

>>> Pl pl (x,y) pM pF = "pl x y (pM*'+' ++ pF*'!')"

estadoInicial :: Int -> Int -> [Picture] -> Estado

O estado inicial do jogo.

gera_players :: Int -> [Player]

desenhaEstado :: Estado -> Picture

Função que desenha o jogo.

A razão pela qual temos uma condição, é porque caso os players morram todos, o jogo acabará, apresentando umas imagens a festejar a vitória.

auxiliar_desenho :: [(Int, Int)] -> [(Int, Int, Int)] -> [(Int, Int)] -> [(Int, Int)] -> [(Int, Int)] -> [(Int, Int)] -> [(Int, Int)] -> [Picture] -> Picture

Auxiliar da função desenhaEstado, onde se desenha cada parte independente no ecrã.

p = desenha as paredes

pl = desenha os players

tij = desenha os tijolos (blocos que são removíveis)

bb = desenha as bombas

pM = desenha os powerMais (aumenta a capacidade de bombas)

expl = desenha as explosões

back = desenha o background

pF = desenha os powerFlames (aumenta o raio das explosões)

desenha_players :: [(Int, Int, Int)] -> [Picture] -> [Picture]

Função que desenha os players. Tem de se ter uma diferente da outra desenha, porque os players estão no mapa com um tipo de dados incompativeis com os outros.

desenha :: [(Int, Int)] -> Picture -> [Picture]

Desenha as imagens no ecra, recebendo um par de ordenadas (correspondes a um eixo imaginário no mapa)

reageEvento :: Event -> Estado -> Estado

Função que altera o estado do jogo quando acontece um evento. Importar a tarefa2.hs

reageTempo :: Float -> Estado -> Estado

Função que altera o estado do jogo quando o tempo avança n segundos. Importamos as funções da Tarefa4.hs para fazer avançar o tempo.

Nota

O bot apenas aparece com 3 players em jogo. Como tem um bug, implementamos no jogo apenas em algumas circunstancias (neste caso precisa de ter 3 players em jogo). Precisamos corrigir o bug.

fr :: Int

Frame rate

dm :: Display

Display mode

main :: IO ()

Função principal que invoca o jogo.

verificaBot :: [Player] -> Bool

Verifica se o bot está ainda em jogo.

coords_powersM :: [PowerMais] -> [(Int, Int)]

Retira do tipo PowerMais, as coordenadas correspondentes a um powerUpMais.

Exemplo

>>> (PM (x,y)) = (x,y)

coords_powersF :: [PowerFlame] -> [(Int, Int)]

Retira do tipo PowerFlame, as coordenadas correspondentes a um powerUpFlame.

Exemplo

>>> (PF (x,y)) = (x,y)

coords_bombas :: [Bomba] -> [(Int, Int)]

Retira do tipo Bomba, as coordenadas correspondentes a um bomba.

Exemplo

>>> (B pl (x,y) r tmp) = (x,y)

coords_players :: [Player] -> [(Int, Int, Int)]

Retira do tipo PLayer, as coordenadas correspondentes a um player.

Exemplo

>>> (Pl pl (x,y) pM pF) = (x,y)

coords_bombasToScale :: [(Int, Int)] -> [(Int, Int)]

Transforma uma coordenada do tipo 1,2,3,4... em pixeis

Exemplo

>>> (1,1) = (334,-334)
>>> (11,1) = (326,334)
>>> (5,8) = (-70,-128)

aux_coords_bombasToScale1 :: [(Int, Int)] -> Int -> (Int, Int) -> [(Int, Int)]

aux_coords_bombasToScale2 :: (Int, Int) -> Int -> Int -> (Int, Int) -> (Int, Int)

coords_blocos_importantes :: [String] -> Char -> [(Int, Int)]

Faz o mesmo que coords_bombasToScale, mas para um determinado char, e usando strings em vez de coodenadas.

aux_coords_blocos_importantes1 :: [String] -> Char -> Int -> [(Int, Int)]

aux_coords_blocos_importantes2 :: String -> Char -> Int -> Int -> [(Int, Int)]

myScaleYourScale :: [(Int, Int, Int)] -> [(Int, Int, Int)]

Faz o mesmo que coords_bombasToScale, mas com um tipo ligeiramente diferente.

myScaleYourScale_aux1 :: [(Int, Int, Int)] -> Int -> (Int, Int) -> [(Int, Int, Int)]

myScaleYourScale_aux2 :: (Int, Int, Int) -> Int -> Int -> (Int, Int) -> (Int, Int, Int)

blocos_inuteis :: [(Int, Int)] -> [String] -> [(Int, Int)]

Filtra coordenadas, de modo a que apenas apresente aquelas que não correspondem ao bloco '#' no mapa.

expl_bombas :: [Bomba] -> [Bomba]

Filtra as bombas (tipo Bomba), apresentando apenas aquelas que têm temporizador 1

getMaybe :: Maybe Char -> Char

Retira o valor de um Data Maybe.

Exemplo

>>> Just x = x
>>> Nothing = ' '