Main

Description

Esta segunda tarefa permite a execução de determinados comandos que movem o jogador (U,D,L,R) e colocação de bombas (B), apresentado o estado final do jogo depois da ação. Usa funções recursivas definidas por nós, mas também algumas ja existentes no Haskell.

Documentation

main :: IO ()

move :: [String] -> Int -> Char -> [String]

mapa -> estado atual do jogo

pl -> nº do jogador a utilizar

chr -> character da ação (comando)

novo_mapa -> mapa com o movimento efectuado

list -> uma lista de elementos

pU -> powerUp

coordenadasDoJogador -> lista com todas as informações da posição do jogador e seus respetivos powerUps

Apresenta o seguinte estado do jogo, conforme a ação (chr)

Syntax --> move mapa pl chr = . . . 

player_no_mapa :: [String] -> Int -> Bool

Verifica se o player pl introduzido existe no mapa

Syntax --> player_no_mapa mapa pl = . . . 

final :: [String] -> Int -> Char -> [String]

Adiciona possiveis powerUps apanhados pelo player pl, ou seja, averigua se existe powerUps na nova posição (atraves da função fake_positionB e fake_positionF) e caso existam adicioná los á string do player (usando a adicionaP para adicionar o powerUp e a remove para retirar a string do powerUp do mapa)

Syntax --> final novo_mapa pl chr = . . . 

fake_positionB :: [String] -> Int -> String

Cria uma string a imitar uma string de powerUps, neste caso de aumento do nº de bombas, usando a nova posição do player pl , para comparar na função final

Syntax --> fake_positionB novo_mapa pl = . . .  

fake_positionF :: [String] -> Int -> String

Cria uma string a imitar uma string de powerUps, neste caso de aumento do raio da bomba, usando a nova posição do player pl, para comparar na função final

Syntax --> fake_positionF novo_mapa pl = . . .  

remove :: [String] -> String -> [String]

Remove de um mapa uma determinada linha

Neste caso foi apenas usada para retirar do novo_mapa a string do powerUp apanhado pelo player pl

Syntax --> remove mapa list = . . . 

Nota

Nest caso, list foi subsituida pelo resultado de fake_positionF||B
para retirar do mapa a string com o powerUp

adicionaP :: String -> [String] -> Int -> [String]

Adiciona ao player pl de um mapa um powerUp, substituindo (usando a função replace) no mapa uma nova string do player pl

Syntax --> adicionaP "+"||"!" mapa pl = . . . 

adiciona_aux :: String -> [String] -> String

Auxiliar de adicionaP

Nota

>>> words "2 39 12 ++"=["2","39","12","++"]
>>> ["2","39","12","++"]==(n:a:b:c:tail) onde n=="2", a=="39", b=="12", c=="++"
>>> unwords ["2","39","12","++"]="2 39 12 ++"
Sendo uma "antiga String" de pl, o words desta pode ter duas formas:
    
    1. length==3, ou seja, nao possui nenhum powerUp, logo pode se
       adicionar diretamente a seguir ao unwords desta.

    2. length==4, ou seja, possui alguns powerUp, logo tem de se ordenar os powerUp
       existentes no player mais o novo (usando a função ordena_powerUps) 

Syntax --> adiciona_aux pU player = . . . 

ordena_powerUps :: String -> String -> String

Ordena os powerUps a serem colocados na String do player _pl_

Syntax --> ordena_powerUps player pU = . . .

mudanca :: [String] -> Int -> Char -> [String]

Função intermédia para fazer o mapa final

mov_valida :: [String] -> Int -> Char -> Bool

Averigua se o movimento apresentado é válido (apenas pode se mover se for um ' ')

Syntax --> move_valida mapa pl chr = . . . 

pos :: [String] -> Int -> [String]

Retira do mapa a String com os dados de um determinado player pl

Syntax --> pos mapa pl = . . . 

Nota

Usamos a words para ser mais facil usar as coordenadas do player pl

>>> words "1 45 790 ++" = ["1","45","790","++"] 

nova_pos :: [String] -> Int -> Char -> [String]

Substitui num mapa, a String do player pl pela nova String deste (nova posição e powerUps)

Syntax --> nova_pos__ mapa pl chr = . . .

replace :: [String] -> String -> [String]

Substitui no mapa, uma String por outra

Syntax --> replace mapa pos_final = . . .

Nota

pos_final é a nova String

pos_final :: [String] -> Int -> Char -> String

Cria uma string nova para o player pl com a sua posição nova depois do movimento

Syntx --> pos_final mapa pl chr = . . .

stringToTuplo :: [String] -> (Int, Int)

Pega numa lista (resuldado de words da String do player pl/) e cria um tuplo, neste caso de coordenadas

Nota

>>> read ("6789")::Int=6789
>>> read ("797a3")::Int=__error__

ex :: [[Char]]

Exemplos para testar o main

ex2 :: [[Char]]

porBomba :: [String] -> Int -> Char -> [String]

Função núcleo que inicia o processo de verificação da possibilidade de realizar a ação por uma bomba

Syntax --> porBomba mapa pl chr = . . .

posBomba :: [String] -> Char -> [String]

Função muito importante no estudo da possibilidade de colocar uma bomba. A partir do estado do mapa e fornecendo o char * a função posBomba devolve uma lista de Strings com todas as coordenadas de bombas presentes no jogo

Syntax --> posBomba mapa chr = . . . 

Nota

Dando um mapa a função devolve algo como: ["* 7 7 1 1 10","* 4 5 2 1 10","..."]

verificaBomba :: [String] -> Int -> [String]

Chama todas as funções que verificam os parâmetros necessários á colocação de uma bomba. Se a função verificaBomba2 se verificar, a função aplica uma função definida por nós: ordenaBomba, que ordena a coordenada da nova bomba a ser colocada no mapa.

Syntax --> verificaBomba mapa pl = . . . 

verificaBomba2 :: Int -> Int -> Bool

Faz uma comparação entre as bombas que o jogador tem colocadas no mapa e os seus powerUps + Se b = número de bombas no mapa do player pl, for maior ou igual a n = numero de powerUps + que o jogador tem, então esse jogador não pode colocar bombas no mapa.

Syntax --> verificaBomba2 b n = . . .

criarCoordBomba :: [String] -> String

Cria coordenadas da nova bomba a ser colocada no mapa. Recebe as coordenadas do jogador que a deseja colocar sobre a forma de [String], devido á utilização da função pré-definida "words"

Syntax --> criarCoordBomba coordenadasDoJogador = . . .

raioBomba :: Int -> String

Calcula o raio de explosáo da nova bomba

Utiliza a função pré-definida show para devolver uma String para ser utilizada na função criarCoordBomba

Função utilizada em conjunto com a contarPowerupsExclama para somar o inteiro que devolve ao _raio de uma bomba normal__ (==1)

Syntax --> raioBomba n = . . .

ordenaBomba :: [String] -> String -> [String]

Ordena as coordenadas da nova bomba no estado atual de um mapa.

Syntax --> ordenaBomba mapa bomba = . . . 

podeBomba :: [String] -> Int -> Bool

Recorre á função compara para poder ser feita a comparação entre as coordenadas onde o jogador se encontra e as coordenadas das bombas já colocadas

Syntax --> mapa pl = . . . 

compara :: [String] -> [String] -> Bool

Compara as coordenadas da posição de um certo player _pl__ com as coordenadas de bombas já colocadas no mapa, se estas coincidirem, então o jogador não pode colocar uma bomba naquele local pois ja se encontra lá uma.

Se a função retribuir True, então pode ser colocada lá uma bomba

Syntax --> compara coordenadasDoJogador bomba = . . .

Nota

>>> compara ["2","2"] ["* 2 2 1 1 10","* 7 7 1 1 10"] = False

contarPowerupsMais :: [String] -> Int

Função que tem como objetivo devolver um número inteiro, que está associado ao número de powerUps + que um certo jogador possui.

Syntax --> contarPowerupsMais  coordenadasDoJogador = . . .

Nota

Recorre á função auxiliar contarMais que pega na linha do mapa (String) e verifica se esta possui os char +, se possuír adiciona 
sempre "+1" ao resultado final, dando como resultado o número de powerUps +.
No fim é sempre adicionado 1 pois o jogador pode sempre por mais uma bomba so que os powerUps que tem.

>>> contarMais "1 5 5 +++" = 4 

contarMais :: String -> Int

contarPowerupsExclama :: [String] -> Int

Função muito idêntica à contarPowerUpsMais, pois segue o mesmo mecanismo, mas para powerUps !

Syntax --> contarPowerupsExclama coordenadasDoJogador = . . .

contarExclama :: String -> Int

Função auxiliar da contarPowerupsExclama, que serve como método de contagem, avaliando cada lista criada a partir da função words das coordenadas do jogador.

Se reconhecer um caratér ! adiciona sempre mais um ao resultado.

Syntax --> contarExclama lista = . . .

bombasNoMapa :: [String] -> Int -> Int

Bombas no Mapa de um certo jogador

Recebe uma lista que contem as coordenadas e informação de todas as bombas. Verifica quantas bombas do jogador pl é que existem no mapa. O sexto elemento de uma coordenada de uma bomba é sempre o número do jogador que a colocou, logo, utilizamos a função pré-definida (!!) para recolhermos o sexto elemento dessa lista, podendo assim compará-lo ao número do jogador que estamos a estudar. No fim, devolve um inteiro que representa o número de bombas de um jogador no mapa.

auxBombas :: String -> Int -> Int