Tarefa6_li1g158

Description

Nesta tarefa, criamos um bot para inteligente para jogar o jogo.

Documentation

type Coord = (Int, Int)

Tipo definido para apresentar Coordenadas.

bot :: [String] -> Int -> Int -> Maybe Char

A função bot recebe o mapa o número do jogador e o tempo de jogo. Primeiro avalia o tempo de jogo comparando-o á fórmula (n-2)^2. Se o tempo for maior então chama a função moves. Se o tempo for menor ou igual á fórmula então chama a função moves_final.

Syntax --> bot mapa pl tempo = . . . 

moves_final :: [String] -> Int -> Maybe Char

Avalia a posição do jogador e se as suas coordenadas pertencerem as coordenadas que vão ser afetadas pela bomba no mapa então é chamada a função descobre_proximo_passo. Caso contrário são comparadas as sua coordenadas á dimensão do mapa e atribuídos comandos para iniciar movimento. Esta função serve para a altura do jogo em que o tempo é menor ou igual a (n-2)^2.

Syntax --> pos mapa pl = . . . 

moves :: [String] -> Int -> Maybe Char

Função que decide se irá ser utilizada uma função de jogada mais defensiva (descobre_prox_passo) ou uma função mais ofensiva (ofensivo) na descoberta do próximo passo do bot. O parâmetro que está presente é o facto de o jogador estar ou não em uma posição que será ou não afetada por uma bomba. Esta função é utilizada enquanto o tempo de jogo é maior que (n-2)^2.

Syntax --> moves mapa pl = . . . 

descobre_prox_passo :: [String] -> Coord -> [Coord] -> [Coord] -> Maybe Char

Função que trata de descobrir qual vai ser o próximo passo que o jogador vai efetuar. Verifica se certas coordenadas pertencem a coordenadas de bombas, ou bombas que vão explodir, ou se são paredes, e dá o comando ao bot.

A função pode ser descrita como um sistema defensivo onde o jogador apenas "foge" aos problemas.

ofensivo :: [String] -> Coord -> Maybe Char

A função ofensivo avalia as coordenadas que estao nas "redondezas" do bot e verifica se alguma destas se trata de um elemento destrutível. Caso se confirme então o bot irá colocar bombas, caso contrário chama a função descobre_prox_passo. Resumidamente, avalia as condições para ver se é benéfico colocar uma bomba ou não.

posBombas :: [String] -> [Coord]

Função que nos dá a posição de todas as bombas em coordenadas no mapa. Recebe um estado do jogo e devolve as coordenadas em que se encontram as bombas.

Syntax --> posBombas mapa = . . . 

posPlayer :: [String] -> Int -> Coord

Função que nos dá a posição de um jogador em coordenadas no mapa. Recebe um estado do jogo e devolve a coordenada em que se encontra o jogador.

Syntax --> posPlayer mapa pl = . . .